Made a start on writing the console - taking a different approach with it
Made a start on writing the console. Got rid of the old console command placeholder implementation...

Instead, I'm going with an approach that allows you to register an arbitrary function/method/functor/lambda and have the arguments figured out automatically via some template magic.

I've got a basic test working that can accept a lambda and call it with a tuple of arguments...

The idea is to generate a tuple at runtime from the parsed arguments and invoke it...